requirements.txt
tensorflow==2.20.0
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
streamlit
opencv-python
flask
